backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-groq
fastapi
uvicorn
python-dotenv
tavily-python
pydantic